+++ include/Elpida/CommonTasks/ReadFile/ReadFile.hpp
@@ -46,7 +46,7 @@ namespace Elpida
 	protected:
 		void prepareImpl() override;
 		TaskOutput finalizeAndGetOutputData() override;
-		size_t calculateTaskResultValue(const Duration& taskElapsedTime) const override;
+		double calculateTaskResultValue(const Duration& taskElapsedTime) const override;
 	private:
 		MemoryFile _file;
 		std::string _filePath;
