+++ include/Elpida/Engine/Task/MultiThreadTask.hpp
@@ -24,7 +24,7 @@ namespace Elpida
 	protected:
 		void prepareImpl() override;
 		TaskOutput finalizeAndGetOutputData() override;
-		virtual size_t calculateTaskResultValue(const Duration& taskElapsedTime) const override;
+		virtual double calculateTaskResultValue(const Duration& taskElapsedTime) const override;
 	private:
 		std::mutex _mutex;
 		std::condition_variable _wakeNotifier;
